Add files via upload
This CNN uses pre-trained GloVe embedding and consists convolutional layers consisting 3 different kernel size and a 300-to-2 linear hidden layer with log_softmax normalization. The input of the CNN is the vector containing the difference between the two given question. The difference of the convolutional outputs for the two vector is calculated and passed into the linear layer.